separate upgrade flow
